requirements-ubuntu.txt
absl-py==1.4.0
aiocache==0.12.0
aioredis==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
codecarbon==2.1.4
contourpy==1.0.7
cryptography==39.0.0
cycler==0.11.0
dill==0.3.6
dm-tree==0.1.8
energyusage==0.0.14
etils==1.0.0
fastapi==0.92.0
fastapi-framework==1.5.3.3
flatbuffers==23.1.21
flwr==1.3.0
fonttools==4.38.0
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.58.0
grpcio==1.46.3
grpcio-tools==1.51.3
h11==0.14.0
h5py==3.6.0
hiredis==2.2.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
iterators==0.0.2
joblib==1.2.0
joulescope==1.1.3
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
munkres==1.1.4
numpy==1.23.2
oauthlib==3.2.2
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
passlib==1.7.4
patsy==0.5.3
Pillow==9.4.0
pip==23.1.2
pooch==1.6.0
promise==2.3
protobuf==3.19.4
psutil==5.9.4
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.5
pyjls==0.5.2
pyjoulescope-driver==1.3.2
PyJWT==2.6.0
pymonocypher==3.1.3.1
pynvml==11.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyRAPL==0.2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
setuptools==67.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
statsmodels==0.13.5
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.8.2
tensorflow-estimator==2.10.0
tensorflow-macos==2.10.0
tensorflow-metadata==1.12.0
tensorflow-metal==0.6.0
termcolor==2.2.0
threadpoolctl==3.1.0
toml==0.10.2
torch==1.11.0
torchvision==0.14.1a0
tqdm==4.64.1
typing_extensions==4.4.0
unicodedata2==15.0.0
urllib3==1.26.14
uvicorn==0.20.0
Werkzeug==2.2.2
wheel==0.38.4
wrapt==1.14.1
zipp==3.12.1
